Fixing 3 issues
