+++ src/Components/FileList/FileList.css
@@ -31,6 +31,7 @@
 
 .no-element {
   display: none;
+  margin-top: 50%;
 }
 
 .no-element.active {
